Refactoring s3 logic
