+++ workspaces/utils/packages/mock-aws-s3-v3/package.json
@@ -38,8 +38,7 @@
     "@smithy/types": "^3.3.0",
     "@types/sinon": "^17.0.3",
     "aws-sdk-client-mock": "^4.0.1",
-    "mock-aws-s3": "^4.0.2",
-    "sinon": "^19.0.2"
+    "mock-aws-s3": "^4.0.2"
   },
   "devDependencies": {
     "@goldstack/utils-git": "0.2.12",
